% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_cells.R
\name{as.data.frame.GoogleWorksheetRef}
\alias{as.data.frame.GoogleWorksheetRef}
\title{coerce a worksheet object into a data.frame}
\usage{
\method{as.data.frame}{GoogleWorksheetRef}(x, header = TRUE, trim = FALSE,
  ...)
}
\description{
this is just a wrapper for \code{RGoogleDocs}'s \code{sheetAsMatrix}
function with the exception that the error caused by an empty sheet
in that function results in an empty data.frame here
}

