% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/babble.R
\name{babble}
\alias{babble}
\title{create a persistent list of pronounceable bubble-babble style words}
\usage{
babble(n, syllables = 1, seed = getOption("babble.seed", 1))
}
\description{
babble returns a vector of n words of at least syllables syllables that are
always the same for a given seed value.  the state of the random number
generator is preserved.
}

