% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{cols2minisurvey}
\alias{cols2minisurvey}
\title{extract repeats from survey data frame}
\usage{
cols2minisurvey(dat, colsets, id = "QID")
}
\description{
\code{cols2minisurvey} takes a data.frame and a list of vectors of column
names or indices and return a minisurvey dataframe.
}

