% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_cells.R
\name{createUpdateRequest}
\alias{createUpdateRequest}
\title{create an update request to modify a sheet}
\usage{
createUpdateRequest(df, sheet)
}
\description{
retrieves a cells feed document for the sheet, compares the cells to the
elements of the data.frame, and forms an xml request to implement the
differences
}

