% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drill.R
\name{drill.svq}
\alias{drill.svq}
\title{drill down (cross tabulate a svq)}
\usage{
drill.svq(q, by, strat = NULL, N = NULL, ...)
}
\arguments{
\item{q}{a svq object}

\item{by}{a factor to split \code{q} by (usually a "select one" svq)}

\item{strat}{an optional factor of strata for stratified samples}

\item{N}{the vector of population values for each stratum}
}
\value{
a list of class \code{svqdrill} with the following elements:
\itemize{
 \item{type}{the question type}
}
for "select" types:
\itemize{
 \item{\code{counts}}
 \item{\code{proportion}}
 \item{\code{margin of error}}
}
for "integer" and "decimal" types:
\itemize{\item{\code{means}}}
for "date" types:
\itemize{\item{\code{counts}}}
}
\description{
\code{drill.svq} is the drill implementation for class \code{svq}
}
\details{
The svq is split by the factor \code{by} and cross sectional measurements
are made for for each subgroup.  The measurement made is determined by the
type of the svq.  "select one" questions generate cross tabulations of
responses in each subgroup and overall.  One table will be generated for the
counts, a second for an estimate of the proportion of each subgroup
responding with each choice, and a third for the standard error of each
estimate.  "select all that apply" (select_multi) questions are similar
but the proportions will not, in general, add up to one.  "decimal" and
"integer" questions will generate a table with the mean, the standard
deviation, the standard error (others to be added).  The last column in each
table will give the totals.  "date" questions will
generate a table of frequencies.

Optionally, \code{q} can be a list of \code{svq} objects of the same length
and question type, and \code{drill.svq} will return a list of results.
\code{by} can also be a list of factors, with a similar result, but the
implementation of this is a little rough.
}
\section{Stratified samples}{


If the data was collected using stratified sampling, estimates and standard
errors can be calculated accordingly by setting \code{strat} and \code{N}.
\code{strat} should be a factor of the same length as q that can be used to
split the data into the strata.  N is a vector of length
\code{length(levels(strat))} with the population size corresponding to each
stratum.  If \code{N} is specified as an integer and strat is \code{NULL},
the finite population correction will be applied to the error estimates.
}

