% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{extract.repeat}
\alias{extract.repeat}
\title{get repeat table}
\usage{
extract.repeat(group, dat, make.names = TRUE, id = row.names(dat),
  id.field = "id", ind.field = "index")
}
\description{
get repeat table
}

