% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_cells.R
\name{getCellsFeed}
\alias{getCellsFeed}
\title{retrieve cells doc}
\usage{
getCellsFeed(sheet, extrema = c(NA, NA, c(dim(sheet))),
  get.empty = any(!is.na(extrema)))
}
\arguments{
\item{sheet}{a \code{GoogleWorksheetRef} object}

\item{extrema}{a vector containing the minimum row and column and the
maximum row and column, in that order.  NA values indicate no minimum
or maximum}

\item{get.empty}{whether to request empty cells (see note below)}
}
\value{
an \code{XMLInternalDocument} containing the feed and its attached
entries
}
\description{
gets an xml document from Google containing all the specified entries
in a sheet
}
\section{Note}{

it may be necessary to retrieve empty cells so they can be updated using
the version code in the \code{rel='edit'} link of the entries for those cells
}

