% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_cells.R
\name{getCellVals}
\alias{getCellVals}
\title{extract cell values}
\usage{
getCellVals(x)
}
\arguments{
\item{x}{an xml doc containing the entries in a cells feed}
}
\value{
a dataframe with \code{row}, \code{col}, and \code{InputValue}
columns
}
\description{
pull a dataframe of rows, columns and values from a cells feed
}
\section{Note}{

the rows in the dataframe will coincide with the order of the entries in the
doc and thus with the order of elements in a nodeset created using
\code{getNodeset(x,"//def:entry",fixns(x))}
}

