% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xlsform.R
\name{makeXLSForm}
\alias{makeXLSForm}
\title{create an xlsform from a simplified form specification}
\usage{
makeXLSForm(df, to = NULL, title = NULL, lang = "English")
}
\arguments{
\item{df}{a data.frame or \code{GoogleWorksheetRef} object}

\item{to}{a file to write the resulting form to in xlsx format}
}
\value{
a list of two data frames with the contents of the survey and choices
sheets in an xlsform
}
\description{
\code{complete.xlsform} takes a data.frame in a special human readable
format and converts it to a legal xlsform, creating the choice sheet in the
process.
}

