% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_cells.R
\name{modifySheet}
\alias{modifySheet}
\title{modify the contents of a sheet}
\usage{
modifySheet(mod, sheet, ...)
}
\arguments{
\item{mod}{a function that takes a data frame and returns another, or another
data.frame}

\item{sheet}{a sheet object}

\item{...}{additional parameters to the function}
}
\value{
invisibly, the server's response to the put request

The sheet contents are fetched and coerced into a data.frame, the function
called on the data.frame, and the results posted to the sheet, replacing the
original contents.  If mod is a data.frame, the contents of the sheet are
simply overwritten
}
\description{
takes a sheet object and a function and changes the contents of the sheet
using the function
}

