% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svy.R
\name{svq.old}
\alias{svq.old}
\title{create a svq (survey question) object from the old-style attribute based
svq}
\usage{
svq.old(x)
}
\description{
create a svq (survey question) object from the old-style attribute based
svq
}

