% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drill.R
\name{tbls2xls}
\alias{tbls2xls}
\title{write tables to xlsx}
\usage{
tbls2xls(tbls, filename = NULL, sheet = createSheet(wb, sheetName =
  "tables"), wb = createWorkbook())
}
\arguments{
\item{tbls}{a (nested) list of tables}

\item{filename}{path to write the results to}

\item{sheet}{the sheet object to write the tables to}

\item{wb}{the workbook in which to create or find the sheet}
}
\value{
the resultant \code{workbook} object
}
\description{
\code{tbls2xlsx} takes a nested list of tables and writes them into a sheet
in an xlsx workbook object, optionally saving it to file.
}
\details{
Originally meant to save drill output to xlsx for making charts, this
function traverses a (tree) list of tables and writes them, in the order
encountered, into a single column on a sheet object from package \code{xlsx}.
Each table must be a matrix and is written with a title formed from its
position in the list, or if the attribute \code{title} is set, that is used.
If \code{sheet} is not specified, a new one is created; if wb is not
specified, a new one is created.  If \code{filename} is specified, the
resultant workbook is written to file.
}

