% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xlsform.R
\name{xlsx2form}
\alias{xlsx2form}
\title{convert an xlsx file into an xlsForm}
\usage{
xlsx2form(f, csv = FALSE, path = NULL, title = NULL, lang = "English")
}
\description{
convert an xlsx file into an xlsForm
}

